(PC-35819) after review refacto to be more logic on import
