+++ src/features/home/components/modules/business/OldBusinessModule.tsx
@@ -5,7 +5,7 @@ import styled, { useTheme } from 'styled-components/native'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { getBusinessUrl } from 'features/home/components/modules/business/helpers/getBusinessUrl'
 import { useShouldDisplayBusinessModule } from 'features/home/components/modules/business/helpers/useShouldDisplayBusinessModule'
-import { BusinessModuleProps } from 'features/home/types'
+import { BusinessModuleProps } from 'features/home/components/modules/business/types'
 import { openUrl } from 'features/navigation/helpers/openUrl'
 import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
 import { analytics } from 'libs/analytics/provider'
