+++ src/features/home/components/modules/exclusivity/ExclusivityModule.native.test.tsx
@@ -3,7 +3,7 @@ import { UseQueryResult } from 'react-query'
 
 import { OfferResponseV2 } from 'api/gen'
 import * as excluOfferAPI from 'features/home/api/useExcluOffer'
-import { ExclusivityModuleProps } from 'features/home/types'
+import { ExclusivityModuleProps } from 'features/home/components/modules/exclusivity/types'
 import { offerResponseSnap as mockOffer } from 'features/offer/fixtures/offerResponse'
 import { render, screen } from 'tests/utils'
