+++ src/features/home/components/modules/exclusivity/ExclusivityModule.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import { ExclusivityBanner } from 'features/home/components/modules/exclusivity/ExclusivityBanner'
 import { ExclusivityExternalLink } from 'features/home/components/modules/exclusivity/ExclusivityExternalLink'
 import { ExclusivityOffer } from 'features/home/components/modules/exclusivity/ExclusivityOffer'
-import { ExclusivityModuleProps } from 'features/home/types'
+import { ExclusivityModuleProps } from 'features/home/components/modules/exclusivity/types'
 import { Spacer } from 'ui/theme'
 
 const UnmemoizedExclusivityModule = ({ offerId, url, ...props }: ExclusivityModuleProps) => {
