+++ src/features/home/components/modules/marketing/types.ts
@@ -1,0 +1,11 @@
+import { InternalNavigationProps } from 'ui/components/touchableLink/types'
+
+export type MarketingBlockProps = {
+  navigateTo: InternalNavigationProps['navigateTo']
+  onBeforeNavigate?: () => void
+  accessibilityLabel?: string
+  backgroundImageUrl?: string
+  AttachedCardComponent: React.ReactNode
+  withGradient?: boolean
+  gradientHeight?: string
+}
