+++ src/features/offer/components/MovieCalendar/MovieCalendar.tsx
@@ -6,13 +6,14 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { MovieCalendarBottomBar } from 'features/offer/components/MovieCalendar/components/MovieCalendarBottomBar'
 import { MovieCalendarDay } from 'features/offer/components/MovieCalendar/components/MovieCalendarDay'
-import {
-  handleMovieCalendarScroll,
-  MOVIE_CALENDAR_PADDING,
-} from 'features/offer/components/MoviesScreeningCalendar/utils'
 import { useHorizontalFlatListScroll } from 'ui/hooks/useHorizontalFlatListScroll'
 import { PlaylistArrowButton } from 'ui/Playlist/PlaylistArrowButton'
 import { getSpacing } from 'ui/theme'
+
+import {
+  handleMovieCalendarScroll,
+  MOVIE_CALENDAR_PADDING,
+} from '../MoviesScreeningCalendar/helpers/handleMovieCalendarScroll'
 
 type Props = {
   dates: Date[]
