+++ src/features/offer/components/MoviesScreeningCalendar/helpers/handleMovieCalendarScroll.ts
